requirements.txt
httpx==0.23.3
python-telegram-bot==20.1
redis==4.5.1
python-dotenv==1.0.0
